refactor: extract ProdBuiltinPairEmitter (Phase 4c step 3b start)
Move genMatchPairData and MatchPairDataLoweredValue out of
ProductCaseSirTypeGenerator into a new ProdBuiltinPairEmitter object —
the first per-repr emitter on the Prod side. ProdDispatch.genMatch's
ProdBuiltinPair branch now routes to ProdBuiltinPairEmitter.genMatch.

selectMatchCase stays in ProductCaseSirTypeGenerator (still shared with
genMatchDataList) and is referenced cross-class for now; it will move to
a shared helper once the remaining Prod emitters land.
