+++ scalus-core/shared/src/main/scala/scalus/compiler/sir/lowering/typegens/ProductCaseSirTypeGenerator.scala
@@ -383,138 +383,6 @@ object ProductCaseSirTypeGenerator extends SirTypeUplcGenerator {
         }
     }
 
-    class MatchPairDataLoweredValue(
-        frs: IdentifiableLoweredValue,
-        snd: IdentifiableLoweredValue,
-        scrutinee: IdentifiableLoweredValue,
-        addScrutineeToScope: Boolean,
-        body: LoweredValue,
-        inPos: SIRPosition
-    ) extends ComplexLoweredValue(
-          Set(frs, snd) ++ (if addScrutineeToScope then Set(scrutinee) else Set.empty),
-          scrutinee,
-          body
-        ) {
-
-        override def sirType: SIRType = body.sirType
-
-        override def representation: LoweredValueRepresentation = body.representation
-
-        override def pos: SIRPosition = inPos
-
-        override def termInternal(gctx: TermGenerationContext): Term = {
-            body.termWithNeededVars(gctx)
-        }
-
-        override def docDef(ctx: LoweredValue.PrettyPrintingContext): Doc = {
-            val left = Doc.text("MatchPairData(")
-            val right = Doc.text(")")
-            body.docRef(ctx).bracketBy(left, right)
-        }
-
-    }
-
-    def genMatchPairData(
-        matchData: SIR.Match,
-        loweredScrutinee: LoweredValue,
-        optTargetType: Option[SIRType]
-    )(using
-        lctx: LoweringContext
-    ): LoweredValue = {
-        val (typeParams, constrDecl, typeArgs) = SIRType
-            .collectProd(loweredScrutinee.sirType)
-            .getOrElse(
-              throw LoweringException(
-                s"Expected product type, got ${loweredScrutinee.sirType.show}",
-                matchData.anns.pos
-              )
-            )
-        // val constrDecl = retrieveConstrDecl(loweredScrutinee.sirType, matchData.anns.pos)
-        val myCase = selectMatchCase(matchData, loweredScrutinee, constrDecl)
-        val prevScope = lctx.scope
-        val (matchVal, addMatchValToScope) = loweredScrutinee match
-            case idv: IdentifiableLoweredValue =>
-                (idv, false)
-            case other =>
-                (
-                  lvNewLazyIdVar(
-                    lctx.uniqueVarName("match_pair_data"),
-                    SIRType.List(SIRType.Data.tp),
-                    SumCaseClassRepresentation.SumBuiltinList(SumCaseClassRepresentation.DataData),
-                    other,
-                    matchData.anns.pos
-                  ),
-                  true
-                )
-        val (frsName, sndName) = myCase.pattern match {
-            case SIR.Pattern.Constr(constr, bindings, typeParamsBindinsg) =>
-                (bindings.head, bindings.tail.head)
-            case SIR.Pattern.Const(_) =>
-                throw LoweringException(
-                  s"Constant pattern not supported for product case class ${constrDecl.name}",
-                  matchData.anns.pos
-                )
-            case SIR.Pattern.Wildcard =>
-                ("_unused1", "_unused2")
-        }
-        val argsMapping = constrDecl.typeParams.zip(typeArgs).toMap
-        val frsTp = SIRType.substitute(constrDecl.params.head.tp, argsMapping, Map.empty)
-        val sndTp = SIRType.substitute(constrDecl.params.tail.head.tp, argsMapping, Map.empty)
-        // val SIRType.partitionGround()
-        val frsRepr = lctx.typeGenerator(frsTp).defaultDataRepresentation(frsTp)
-        val sndRepr = lctx.typeGenerator(sndTp).defaultDataRepresentation(sndTp)
-
-        if lctx.targetProtocolVersion >= MajorProtocolVersion.vanRossemPV then {
-            // For PlutusV4: use Case on Pair - frs and snd are lambda parameters
-            val frsVarId = lctx.uniqueVarName(frsName)
-            val frsVar = new VariableLoweredValue(
-              id = frsVarId,
-              name = frsName,
-              sir = SIR.Var(frsVarId, frsTp, AnnotationsDecl.empty),
-              representation = frsRepr,
-              optRhs = None // lambda parameter, not derived from builtin
-            )
-            val sndVarId = lctx.uniqueVarName(sndName)
-            val sndVar = new VariableLoweredValue(
-              id = sndVarId,
-              name = sndName,
-              sir = SIR.Var(sndVarId, sndTp, AnnotationsDecl.empty),
-              representation = sndRepr,
-              optRhs = None // lambda parameter, not derived from builtin
-            )
-            // Add pair element vars to scope so the body can reference them
-            lctx.scope = lctx.scope.addAll(scala.collection.immutable.List(frsVar, sndVar))
-            val lwBody = lctx.lower(myCase.body, optTargetType)
-            lvCasePair(matchVal, frsVar, sndVar, lwBody, matchData.anns.pos)
-        } else {
-            // For V1-V3: use fstPair/sndPair builtins
-            val frs = lvNewLazyNamedVar(
-              frsName,
-              frsTp,
-              frsRepr,
-              lvBuiltinApply(SIRBuiltins.fstPair, matchVal, frsTp, frsRepr, myCase.anns.pos),
-              myCase.anns.pos
-            )
-            val snd = lvNewLazyNamedVar(
-              sndName,
-              sndTp,
-              sndRepr,
-              lvBuiltinApply(SIRBuiltins.sndPair, matchVal, sndTp, sndRepr, myCase.anns.pos),
-              myCase.anns.pos
-            )
-            val lwBody = lctx.lower(myCase.body, optTargetType)
-            MatchPairDataLoweredValue(
-              frs,
-              snd,
-              matchVal,
-              addMatchValToScope,
-              lwBody,
-              matchData.anns.pos
-            )
-        }
-
-    }
-
     def genConstrDataConstr(
         constr: SIR.Constr,
         loweredArgs: Seq[LoweredValue],
